content length to chrome it being considered zero in many cases, even when response put a value greater than zero. Maybe something during the OPTIONS called its not being well and after that chrome start to block everything else. In a fast overview over network in chrome(f12) noticed that the first CALL with options area exactly equal to the original one.
Another point, the old heroku app that you used in video, its not working anymore, so changed to another. And in this new one, the second test it's not working, so for sure it will never works on the proxy.

About the rest, i am having the content lenght problem going to zero for chrome perspective